Added some math functions, refactored the Vector sistem to use templates and fixed warnings

+++ Spyen/src/Renderer/Shader.cpp
@@ -160,7 +160,7 @@ namespace Spyen
 		GLuint location = glGetUniformLocation(m_RendererID, name.c_str());
 		glUniformHandleui64ARB(location, handle);
 	}
-	void Shader::SetUniformHandles(const std::string& name, const uint64_t* handles, size_t count)
+	void Shader::SetUniformHandles(const std::string& name, const uint64_t* handles, uint32_t count)
 	{
 		GLuint location = glGetUniformLocation(m_RendererID, name.c_str());
 		glUniformHandleui64vARB(location, count, handles);
